+++ datahub/investment/opportunity/permissions.py
@@ -1,0 +1,8 @@
+from datahub.core.utils import StrEnum
+
+
+class LargeCapitalOpportunityPermission(StrEnum):
+    """Permission codename constants."""
+
+    view_large_capital_opportunity = 'view_largecapitalopportunity'
+    export = 'export_largecapitalopportunity'