+++ CHANGELOG.md
@@ -1,3 +1,75 @@
+# Data Hub API 36.16.0 (2021-04-16)
+
+
+## Features
+
+- **Advisers** Feature flags can now be applied on a per-user basis through Django Admin.
+
+  - Adds a new UserFeatureFlag model
+  - Adds a new "features" field to the Advisor model
+  - Exposes "active_features" on the `whoami` endpoint
+
+## API
+
+- **Investment** A new `POST /v4/search/large-capital-opportunity/export` endpoint has been added
+  that enables export of search results in CSV format.
+
+
+# Data Hub API 36.15.0 (2021-04-13)
+
+
+## Features
+
+- **Investment** A new search app has been added for Large Capital Opportunity.
+- **Investment** **Investment** Project Proposition has been added to the admin area.
+- Add Trade agreement interaction choice and service update.
+
+## Internal changes
+
+- Trade agreements moved from events to generic metadata in order to add trade agreements to interactions going forward and update trade agreements
+
+## API
+
+- **Investment** A new `POST /v4/search/large-capital-opportunity` endpoint has been added.
+  The endpoint enables the following filters:
+
+  Main filters:
+
+      type
+      status
+      name
+      created_by
+
+  Detail filters:
+
+      uk_region_location
+      promoter
+      promoter_name
+      lead_dit_relationship_manager
+      required_checks_conducted
+      required_checks_conducted_by
+      asset_class
+      opportunity_value_type
+      opportunity_value_start
+      opportunity_value_end
+      construction_risk
+
+  Requirement filters:
+
+      total_investment_sought_start
+      total_investment_sought_end
+      current_investment_secured_start
+      current_investment_secured_end
+      investment_type
+      estimated_return_rate
+      time_horizon
+
+  Extra filters:
+
+      created_on_after
+      created_on_before
+
+
 # Data Hub API 36.14.0 (2021-04-09)